+++ src/id/mod.rs
@@ -49,9 +49,9 @@ where
 }
 #[derive(PartialEq, Debug, Clone, Copy)]
 /// * z: The zoom level, a u16 value. It defines the coordinate space boundaries.
-/// * f: The value for the F (vertical) dimension as a `DimensionRange<i64>`.
-/// * x: The value for the X dimension as a `DimensionRange<u64>`.
-/// * y: The value for the Y dimension as a `DimensionRange<u64>`.
+/// * f: The value for the F (vertical) dimension as a `DimensionRange<i32>`.
+/// * x: The value for the X dimension as a `DimensionRange<u32>`.
+/// * y: The value for the Y dimension as a `DimensionRange<u32>`.
 /// * i: The time interval in seconds, a `u32` value.
 /// * t: The time index value as a `DimensionRange<u32>`.
 ///
@@ -93,9 +93,9 @@ where
 #[derive(Hash, Eq, PartialOrd, Ord)]
 pub struct SpaceTimeId {
     z: u16,
-    f: DimensionRange<i64>,
-    x: DimensionRange<u64>,
-    y: DimensionRange<u64>,
+    f: DimensionRange<i32>,
+    x: DimensionRange<u32>,
+    y: DimensionRange<u32>,
     i: u32,
     t: DimensionRange<u32>,
 }
@@ -141,9 +141,9 @@ impl SpaceTimeId {
     /// - If `t` is not `Any` when `i == 0`
     pub fn new(
         z: u16,
-        f: DimensionRange<i64>,
-        x: DimensionRange<u64>,
-        y: DimensionRange<u64>,
+        f: DimensionRange<i32>,
+        x: DimensionRange<u32>,
+        y: DimensionRange<u32>,
         i: u32,
         t: DimensionRange<u32>,
     ) -> Result<Self, String> {
@@ -155,10 +155,10 @@ impl SpaceTimeId {
         }
 
         fn validate_xy_dim(
-            dim_val: &DimensionRange<u64>,
+            dim_val: &DimensionRange<u32>,
             z: u16,
-        ) -> Result<DimensionRange<u64>, String> {
-            let max = (1u64 << z) - 1;
+        ) -> Result<DimensionRange<u32>, String> {
+            let max = (1u32 << z) - 1;
 
             match *dim_val {
                 Single(v) => {
@@ -232,10 +232,10 @@ impl SpaceTimeId {
         }
 
         fn validate_f_dim(
-            dim_val: &DimensionRange<i64>,
+            dim_val: &DimensionRange<i32>,
             z: u16,
-        ) -> Result<DimensionRange<i64>, String> {
-            let limit = 1i64 << z;
+        ) -> Result<DimensionRange<i32>, String> {
+            let limit = 1i32 << z;
             let max_f = limit - 1;
             let min_f = -limit;
